multi_tool_agent/requirements.txt
# google-generativeai
# google-adk
# python-dotenv

google-generativeai==0.8.3
python-dotenv==1.0.1
flake8==7.1.1